# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Insurance OTP service is used for sending and varifying OTPs",
        "title": "Insurance OTP service",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/otp": {
            "get": {
                "description": "Send a OTP to client by sms",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "otp"
                ],
                "summary": "Send a OTP",
                "operationId": "get-otp-to-phon-number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Phone number to send OTP",
                        "name": "phone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetOtpRespObj"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Send a OTP to client by sms",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "otp"
                ],
                "summary": "Send a OTP",
                "operationId": "post-otp-to-phon-number",
                "parameters": [
                    {
                        "description": "Body should contain phone number, key, and otp",
                        "name": "validate_otp_req_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ValidateOtpReqBody"
                        }
                    }
                ],
                "responses": {
                    "204": {},
                    "400": {
                        "description": "This status is returned if wrong otp has been sent",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "This status is returned if key status is no more valid",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "This status is returned if otp sent in more than 60 sec",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "This status is returned if otp reached 3 attemps of validation",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.GetOtpRespObj": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                }
            }
        },
        "handlers.ValidateOtpReqBody": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "otp": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "helpers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}